Moved code in charge of data storages to new file
Should make it easier to read
